TDQ-13389: add the class AbstractCardinalityStatistics
