+++ dataquality-statistics/src/main/java/org/talend/dataquality/statistics/cardinality/CardinalityHLLStatistics.java
@@ -15,14 +15,12 @@ package org.talend.dataquality.statistics.cardinality;
 import com.clearspring.analytics.stream.cardinality.HyperLogLog;
 
 /**
- * Cardianlity statistics bean of hypper log log .
+ * Cardinality statistics bean of hyper log log .
  * 
  * @author zhao
  *
  */
-public class CardinalityHLLStatistics {
-
-    private long count = 0L;
+public class CardinalityHLLStatistics extends AbstractCardinalityStatistics {
 
     private HyperLogLog hyperLogLog = null;
 
@@ -33,27 +31,12 @@ public class CardinalityHLLStatistics {
         return hyperLogLog;
     }
 
-    public long getDuplicateCount() {
-        return count - getDistinctCount();
-    }
-
-    public long getCount() {
-        return count;
-    }
-
     public long getDistinctCount() {
         return hyperLogLog.cardinality();
-    }
-
-    public void incrementCount() {
-        count += 1;
     }
 
     public void setHyperLogLog(HyperLogLog hyperLogLog2) {
         this.hyperLogLog = hyperLogLog2;
     }
 
-    public void setCount(long count) {
-        this.count = count;
-    }
 }
